backend/requirements.txt
# --- API & Core Framework ---
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
gunicorn==21.2.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv==1.0.0

# --- Security & Auth ---
# Note: Ensure your code imports match these libraries
python-multipart>=0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# --- Azure Services ---
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
azure-ai-projects>=2.0.0b1
azure-core>=1.30.0

# --- Database ---
pyodbc>=5.1.0

# --- Data Processing ---
pandas>=2.2.3
openpyxl>=3.1.5
scikit-learn>=1.4.0

# --- PDF Handling ---
PyPDF2==3.0.1
pdfplumber>=0.10.0